Edit, FillForm, View, ConverionForm, DownloadAsForm moved to actions.py
